author

+++ pc-core/src/main/java/com/zte/msg/pushcenter/pccore/mapper/SecretMapper.java
@@ -16,7 +16,7 @@ public interface SecretMapper extends BaseMapper<SecretKey> {
 
     String selectAppRole(String appKey);
 
-    String selectAppSecret(Integer appId);
+    String selectAppSecret(Long appId);
 
     SecretKeyResDTO selectSecretKey(Integer appId);
 }
